PackageManager: Fix of 'Spoof Google Play Store as installer...'
The original commit to spoof the Google Play Store as installer in case of
AuroraStore was using the 'equals()' method, which crashes with a NullPointer
exception, if the installer string has a Null value. Different from the original
assumption of this value always being present, there are indeed rarely
situations, where this is not the case. The == operator does not crash in such
cases.

Change-Id: I5be015e11feb61bec041be281924b45648b1afb3